% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_packages.R
\docType{data}
\name{dat_packages}
\alias{dat_packages}
\title{List of packages for SCHOL^R}
\format{
6 variables:
\itemize{
\item \strong{order}: Order in which the packages should be installed.
\item \strong{use}: The type of task for which the package is useful.
\item \strong{origin}: Indicate whether the packages should be installed from cran, GitHub, or BioConductor
\item \strong{repository}: If GitHub, the developer.
\item \strong{package}: Name of the package.
\item \strong{version}: Version of the package which should be installed. Latest if not specified.
}
}
\usage{
data("dat_packages")
}
\description{
List of packages constitutive of the SCOL^R environment, with their sources and versions specified if necessary.
}
\author{
Nicolas Mangin
}
\keyword{environment}
\keyword{packages}
