% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Document data for a package}
\usage{
document_data(x = NULL, datname = NULL, path = "R")
}
\arguments{
\item{x}{Dataframe or tibble. Data for which a description should be created.}

\item{datname}{Character. Name of the dataset.}

\item{path}{Character. Path to which the description should be added.}
}
\description{
Function creating a .R file for a dataset.
}
\author{
Nicolas Mangin
}
