% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_scholR.R
\name{install_scholR}
\alias{install_scholR}
\title{Install scholR packages}
\usage{
install_scholR(
  preinstall_requisites = FALSE,
  install_for_education = TRUE,
  install_for_research = FALSE,
  keep_existing = TRUE
)
}
\arguments{
\item{preinstall_requisites}{Logical. Whether packages necessary for all other scholR pachages should be pre-installed.}

\item{install_for_education}{Logical. Whether packages from the education universe should be installed.}

\item{install_for_research}{Logical. Whether packages from the research universe should be installed.}

\item{keep_existing}{Logical. Whether already installed packages should be reinstalled.}
}
\description{
Function installing packages required by, and from, the scholR environment.
}
\author{
Nicolas Mangin
}
