% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_functions}
\alias{list_functions}
\title{Replace a pattern in several documents}
\usage{
list_functions(path = "R/", type = ".R")
}
\arguments{
\item{path}{Character. Relative path to the directory where the files are.}

\item{type}{Character. File extension to scan.}
}
\description{
List all the packages used in a set of .R files (identified with ::)
}
\author{
Nicolas Mangin
}
