% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_used_packages_and_functions.R
\name{list_used_packages_and_functions}
\alias{list_used_packages_and_functions}
\title{List packages and functions}
\usage{
list_used_packages_and_functions(path = "R")
}
\arguments{
\item{path}{Character. Path to the folder containing the files.}
}
\value{
Tibble with packages and function names, as well as the related imports.
}
\description{
Function listing all the packages and functions used in a folder specified by a path.
}
\author{
Nicolas Mangin
}
