% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_files_containing_pattern.R
\name{open_files_containing_pattern}
\alias{open_files_containing_pattern}
\title{Open files containing a pattern}
\usage{
open_files_containing_pattern(path, pattern)
}
\arguments{
\item{path}{Character. Path to the folder containing the files.}

\item{pattern}{Character. Regex pattern to be replaced.}
}
\description{
Function identifying and opening files containing the specified regex pattern.
}
\author{
Nicolas Mangin
}
