% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_environment.R
\name{prepare_environment}
\alias{prepare_environment}
\title{Install scholR environment}
\usage{
prepare_environment(
  cran = TRUE,
  github = TRUE,
  bioconductor = TRUE,
  hugo = FALSE,
  webshot = FALSE
)
}
\arguments{
\item{cran}{Logical. TRUE to install cran packages.}

\item{github}{Logical. TRUE to install github packages.}

\item{bioconductor}{Logical. TRUE to install BioConductor packages.}

\item{hugo}{Logical. TRUE to install.}

\item{webshot}{Logical. TRUE to install.}
}
\description{
Install all packages supporting the scholR universe in the appropriate versions.
}
\author{
Nicolas Mangin
}
