% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_set_paths.R
\name{project_set_paths}
\alias{project_set_paths}
\title{Complete course folder paths}
\usage{
project_set_paths(project_folder)
}
\arguments{
\item{project_folder}{Character. Course folder as selected by the user.}
}
\value{
A list of folder paths based on a specified course structure and used by the application.
}
\description{
Retrieve a list of paths to relevant databases or subfolders in the selected course folder on the local computer.
}
\author{
Nicolas Mangin
}
