% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\docType{data}
\name{references}
\alias{references}
\title{References}
\format{
A data frame with 8 rows and 26 variables
\describe{
\item{bibtype}{}
\item{key}{}
\item{author}{}
\item{title}{}
\item{journal}{}
\item{year}{}
\item{month}{}
\item{volume}{}
\item{number}{}
\item{pages}{}
\item{publisher}{}
\item{booktitle}{}
\item{editor}{}
\item{institution}{}
\item{school}{}
\item{address}{}
\item{edition}{}
\item{note}{}
\item{doi}{}
\item{url}{}
\item{abstract}{}
\item{keywords}{}
\item{isbn}{}
\item{issn}{}
\item{jnl}{}
\item{field}{}
}
}
\usage{
references
}
\description{
References to the bibliogR package.
}
\keyword{datasets}
