% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_in_files.R
\name{replace_in_files}
\alias{replace_in_files}
\title{Replace a pattern in several documents}
\usage{
replace_in_files(path, pattern, replacement, type = ".Rmd", folders = FALSE)
}
\arguments{
\item{path}{Character. Relative (from the working directory) or absolute path to the files.}

\item{pattern}{Character. Pattern to find and replace in each document.}

\item{replacement}{Character. Pattern to put instead.}

\item{type}{Character. Extension of the files to be processed}

\item{folders}{Logical. Whether documents in sub-folders should be included.}
}
\value{
Replace in the same folder the documents where the string was found by the same documents where the string was replaced. The change is destructive.
}
\description{
Replace the specified pattern of characters by another in all the documents within a path.
}
\author{
Nicolas Mangin
}
