% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_pattern_in_files.R
\name{replace_pattern_in_files}
\alias{replace_pattern_in_files}
\title{Replace pattern in files}
\usage{
replace_pattern_in_files(path, pattern, replacement)
}
\arguments{
\item{path}{Character. Path to the folder containing the files.}

\item{pattern}{Character. Regex pattern to be replaced.}

\item{replacement}{Chatacter. Replacement for the regex pattern.}
}
\description{
Function identifying files containing the specified regex pattern and replacing this pattern by the desired replacement.
}
\author{
Nicolas Mangin
}
