% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\docType{data}
\name{required_packages}
\alias{required_packages}
\title{Required packages}
\format{
A data frame with 39 rows and 2 variables
\describe{
\item{Package}{}
\item{Version}{}
}
}
\usage{
required_packages
}
\description{
List of packages used in the scholR set of packages.
}
\keyword{datasets}
