% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlecase.R
\name{titlecase}
\alias{titlecase}
\title{Title Case Conversion Functions}
\description{
Functions to convert text to title case following the New York Times
Manual of Style conventions, and to process BibTeX files.
}
